backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
reportlab==4.0.7
langchain==0.1.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.25.0
numpy>=1.24.0
python-dotenv==1.0.0
google-generativeai>=0.3.0